Rename files, complete trip-main
