+++ public/index.html
@@ -17,15 +17,13 @@
         <div class="trip-main">
           <!-- Маршрут и стоимость -->
 
-          <div class="trip-main__trip-controls  trip-controls">
-            <h2 class="visually-hidden">Switch trip view</h2>
+          <!-- Обертка -->
+          
             <!-- Меню -->
 
-            <h2 class="visually-hidden">Filter events</h2>
             <!-- Фильтры -->
-          </div>
 
-          <button class="trip-main__event-add-btn  btn  btn--big  btn--yellow" type="button">New event</button>
+          <!-- Кнопка New Event -->
         </div>
       </div>
     </header>
