+++ src/view/site-controls/site-controls-menu.js
@@ -4,7 +4,7 @@ const createMenuTemplate = () => {
       <a class="trip-tabs__btn" href="#">Table</a>
       <a class="trip-tabs__btn  trip-tabs__btn--active" href="#">Stats</a>
     </nav>
-    `;
+  `;
 };
 
 export {
