+++ src/view/site-info/site-info-main.js
